Flow: cleanup flowtypes
